Move parse functions in the parser directory
