+++ Parser/Token.hpp
@@ -1,24 +1,23 @@
-#ifndef TOKEN_HPP
-#define TOKEN_HPP
+#ifndef PARSER_TOKEN_HPP
+#define PARSER_TOKEN_HPP
 
 #include <string>
 
-#include "Webserv.hpp"
+#include "../Webserv.hpp"
 
 class Token {
    private:
 	e_TokenType _type;
 	std::string _lexeme;
 	int _line;
-	int	_colon;
-	
+	int _colon;
 
    public:
 	Token(e_TokenType type, std::string lexeme, int line, int colon);
 	std::string getAll();
 	t_TokenType getType();
 	int getLine();
-	int	getColon();
+	int getColon();
 	int getLength();
 };
