+++ main.cpp
@@ -1,4 +1,4 @@
-#include "Parser.hpp"
+#include "Parser/Parser.hpp"
 
 int main(int argc, char** argv) {
 	std::string fileName;
